requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Document processing
pymupdf==1.26.4
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2

# PageIndex dependencies
openai==1.101.0
tiktoken==0.11.0
PyPDF2==3.0.1

# Utils
python-dotenv==1.1.0
pydantic==2.10.0
pydantic-settings==2.6.0
pyyaml==6.0.2
